🎨update code structure with reference to code review
